Update build_pulseaudio-modules-droid_package.sh
If local git repo exists do a pull instead of clone.
Added check if rpm packages were build successfully.

+++ build_pulseaudio-modules-droid_package.sh
@@ -9,13 +9,24 @@ PKG=pulseaudio-modules-droid
 SPEC=$PKG
 
 cd $MER_ROOT/devel/mer-hybris
-git clone https://github.com/mer-hybris/$PKG.git
+
+if [ ! -d $PKG ]; then
+	git clone https://github.com/mer-hybris/$PKG.git
+else
+	git pull
+fi
+
 cd $PKG
 mb2 -s rpm/$SPEC.spec -t $VENDOR-$DEVICE-armv7hl build
+
+if [ $(ls RPMS/*.rpm | wc -l) -eq 0 ]; then
+	echo "Error: No RPMs found for $PKG"
+	exit 1
+fi
+
 mkdir -p $ANDROID_ROOT/droid-local-repo/$DEVICE/$PKG/
 rm -f $ANDROID_ROOT/droid-local-repo/$DEVICE/$PKG/*.rpm
 mv RPMS/*.rpm $ANDROID_ROOT/droid-local-repo/$DEVICE/$PKG
 
 createrepo $ANDROID_ROOT/droid-local-repo/$DEVICE
 sb2 -t $VENDOR-$DEVICE-armv7hl -R -msdk-install zypper ref
-
